requirements.txt
asgiref==3.8.1
asttokens==3.0.0
boto3==1.36.21
botocore==1.36.21
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.1.4
django-extensions==3.2.3
django-storages==1.14.5
executing==2.1.0
Faker==33.3.1
graphviz==0.20.3
icecream==2.1.3
idna==3.10
ipython==8.31.0
jedi==0.19.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
matplotlib-inline==0.1.7
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydotplus==2.0.2
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
s3transfer==0.11.2
scipy==1.15.1
shell==1.0.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
visdom==0.2.4
wcwidth==0.2.13
websocket-client==1.8.0
